move autoplay video block to streams app
