+++ streams/blocks.py
@@ -1,4 +1,37 @@
+from django.forms.utils import flatatt
+from django.utils.html import format_html, format_html_join
 from wagtail.core import blocks
+from wagtailmedia.blocks import AbstractMediaChooserBlock
+
+
+class VideoBlock(AbstractMediaChooserBlock):
+    def render_basic(self, value, context=None):
+        if not value:
+            return ''
+
+        player_code = '''
+        <div>
+            <video width="320" height="240" controls>
+                {0}
+                Your browser does not support the video tag.
+            </video>
+        </div>
+        '''
+
+        return format_html(player_code, format_html_join(
+            '\n', "<source{0}",
+            [[flatatt(s)] for s in value.sources]
+        ))
+
+
+class AutoPlayVideoBlock(blocks.StructBlock):
+    video = VideoBlock(required=False)
+    caption = blocks.CharBlock(required=False)
+
+    class Meta:
+        icon = 'media'
+        label = 'Autoplay Video'
+        template = 'streams/autoplay_video_block.html'
 
 
 class ParagraphBlock(blocks.RichTextBlock):
